add a pass to replace constant division into multiplication

+++ scripts/run_FBA.sh
@@ -1,22 +1,23 @@
 #!/bin/bash
 
-if [ "$#" -ne 6 ]; then
+if [ "$#" -ne 5 ]; then
   echo "Usage: $0 <NAF> <source> <precision> <min> <max>"
   exit 1
 fi
 
-DIR=./temp/
-if [ ! -d $DIR ]; then
-    mkdir temp
-fi
+# Build NAF list
+declare -A NAF
+NAF[tanh]='tanh(x)'
+NAF[gelu]='x/2*(1+erf(x/sqrt(2)))'
+NAF[erf]='erf(x)'
 
 # <source> = initial FBA candidate IRB
 
-# 1. build sub-func table and filter candidates of PAG (+ OP-Analyzer)
-subfuncTbl=($(/usr/local/bin/opt -load-pass-plugin ./build/lib/libOpAnalyzer.so -passes=analyze-op,dce -S -disable-output $2 2>&1))
+# 0. Op-Analyzer (1) replace constant fdiv into fmul
+/usr/local/bin/opt -load-pass-plugin ./build/lib/libReplaceDivtoMul.so -passes=replace-div-mul,dce -S $2 -o temp/temp_$1.ll
 
-# copy temporal $2
-cp $2 temp/temp_$1.ll
+# 1. Op-Analyzer (2) build sub-func table and filter candidates of PAG
+subfuncTbl=($(/usr/local/bin/opt -load-pass-plugin ./build/lib/libOpAnalyzer.so -passes=analyze-op,dce -S -disable-output temp/temp_$1.ll 2>&1))
 
 # TODO: f order by heuristic cost estimation
 for f in "${subfuncTbl[@]}"; do
@@ -25,13 +26,13 @@ for f in "${subfuncTbl[@]}"; do
   /usr/local/bin/llc -filetype=obj temp/temp_$f.ll -o temp_$f.o
   /usr/local/bin/clang++ ./scripts/checkMax.cpp temp_$f.o -o checkMax -lm
   fMax=$(./checkMax $4 $5)
-  echo "($4, $5) fMax : $fMax"
+  echo "function $f, ($4, $5) fMax : $fMax"
 
   found=0
-  for deg in $(seq 11 2 27); do
+  for deg in $(seq 11 1 27); do
     # 2. run PAG with desired input range of subfunction
     # TODO: modify lolremez output (function) name
-    lolremez --double -r "-$fMax:$fMax" "$f(x)" -d $deg > temp/temp_$f.c
+    lolremez --double -r "-$fMax:$fMax" "${NAF[$f]}" -d $deg > temp/temp_$f.c
 
     # 3. compile PA
     /usr/local/bin/clang -O2 -c -emit-llvm temp/temp_$f.c -o temp/temp_$f.bc
@@ -44,7 +45,7 @@ for f in "${subfuncTbl[@]}"; do
 
     # 5. check the validity of current PA for sub-func (UDC Tracker)
     tmp=$(bash ./scripts/checkValid.sh replaced $3 $4 $5)
-    fRate=$(echo "$tmp" | awk '{print $3}' | tr -d '%')
+    fRate=$(echo "$tmp" | awk '{print $4}' | tr -d '%')
     if [[ "$fRate" == "0" ]]; then
         echo "Minimum approximation degree of $f = $deg"
         found=1
@@ -70,31 +71,38 @@ done
 # - NAF is replaced with approximated function
 # - The function returns the input of div
 /usr/local/bin/opt -load-pass-plugin ./build/lib/libGetDivisorRange.so -passes=get-div-range,dce -S -o temp/temp_$1_div.ll temp/temp_$1.ll
-/usr/local/bin/llc -filetype=obj temp/temp_$1_div.ll -o temp_$1_div.o
-/usr/local/bin/clang++ ./scripts/checkMax.cpp temp_$1_div.o -o checkMax -lm
-divMax=$(./checkMax $4 $5)
-echo "divMax : $divMax"
 
-# 6. compile approximated divide (inverse)
-/usr/local/bin/clang -O2 -c -emit-llvm math/_inverse.c -o temp/_inverse.bc
-/usr/local/bin/llvm-dis temp/_inverse.bc
-/usr/local/bin/opt -disable-output temp/_inverse.ll
+isExistDiv=$(/usr/local/bin/opt -load-pass-plugin ./build/lib/libGetDivisorRange.so -passes=get-div-range,dce -S -o temp/temp_$1_div.ll temp/temp_$1.ll 2>&1)
 
-# 7. link top function and approximated polynomial
-/usr/local/bin/llvm-link temp/temp_$1.ll temp/_inverse.ll -S -o temp/merged_div.ll
+if [[ $isExistDiv == "1" ]]; then
+  /usr/local/bin/llc -filetype=obj temp/temp_$1_div.ll -o temp_$1_div.o
+  /usr/local/bin/clang++ ./scripts/checkMax.cpp temp_$1_div.o -o checkMax -lm
+  divMax=$(./checkMax $4 $5)
+  echo "divMax : $divMax"
 
-for d in $(seq 11 1 17); do
-  # 8. run ReplaceDiv pass
-  /usr/local/bin/opt -load-pass-plugin ./build/lib/libReplaceDiv.so -passes=replace-div,dce -div-max=$divMax -iter-d=$d -S -o temp/$1_result.ll temp/merged_div.ll
+  # 6. compile approximated divide (inverse)
+  /usr/local/bin/clang -O2 -c -emit-llvm math/_inverse.c -o temp/_inverse.bc
+  /usr/local/bin/llvm-dis temp/_inverse.bc
+  /usr/local/bin/opt -disable-output temp/_inverse.ll
 
-  # 5. check the validity of current approximated polynomial for sub-func
-  tmp=$(bash ./scripts/checkValid.sh $1_result $3 $4 $5)
-  fRate2=$(echo "$tmp" | awk '{print $3}' | tr -d '%')
-  if [[ "$fRate2" == "0" ]]; then
-      echo "Minimum d = $d"
-      cp temp/$1_result.ll results/
-      break
-  fi
-done
+  # 7. link top function and approximated polynomial
+  /usr/local/bin/llvm-link temp/temp_$1.ll temp/_inverse.ll -S -o temp/merged_div.ll
+
+  for d in $(seq 11 1 17); do
+    # 8. run ReplaceDiv pass
+    /usr/local/bin/opt -load-pass-plugin ./build/lib/libReplaceDiv.so -passes=replace-div,dce -div-max=$divMax -iter-d=$d -S -o temp/$1_result.ll temp/merged_div.ll
+
+    # 5. check the validity of current approximated polynomial for sub-func
+    tmp=$(bash ./scripts/checkValid.sh $1_result $3 $4 $5)
+    fRate2=$(echo "$tmp" | awk '{print $4}' | tr -d '%')
+    if [[ "$fRate2" == "0" ]]; then
+        echo "Minimum d = $d"
+        cp temp/$1_result.ll results/
+        break
+    fi
+  done
+else
+  cp temp/temp_$1_div.ll results/$1_result.ll
+fi
 
 rm -rf *.o checkMax
